Completed homework
