+++ Homework-II_joseph_galloway/homework_II.md
@@ -1,6 +1,6 @@
 # **Homework II**
 
-Author: Joseph D. Galloway II  
+Author: Joseph D. Galloway II
 Course: Design Optimization (ME 598)  
 *[Google Colab Notebook](wwww.google.com)*
 
@@ -17,8 +17,10 @@ Course: Design Optimization (ME 598)
 ## Problem II
 ![Part_I](Images/problemII.png)
 ![Part_I](Images/problemII_I.png)
-
-
+![Part_I](Images/problemII_gradient.png)
+![Part_I](Images/problemII_gradient_II.png)
+![Part_I](Images/problemII_newton.png)
+![Part_I](Images/problemII_newton_II.png)
 
 
 ## Problem III
